Fix wrong parameter types on 2 patches

+++ StressandTaxes/Prices/Config_Patch.cs
@@ -3,11 +3,9 @@ using HarmonyLib;
 
 namespace StressandTaxes.Costs.Patches
 {
-	// Token: 0x02000004 RID: 4
 	[HarmonyPatch(typeof(GameConfigData), "HullRepairCostPerSquare", MethodType.Getter)]
 	public static class GameConfigData_HullRepair_Patch
 	{
-		// Token: 0x06000007 RID: 7 RVA: 0x00002C74 File Offset: 0x00000E74
 		[HarmonyPostfix]
 		static void HullRepairCostPerSquare(ref decimal __result)
 		{
@@ -18,9 +16,8 @@ namespace StressandTaxes.Costs.Patches
     [HarmonyPatch(typeof(GameConfigData), "NightSanityModifier", MethodType.Getter)]
     public static class GameConfigData_SanityModifier_Patch
     {
-        // Token: 0x06000007 RID: 7 RVA: 0x00002C74 File Offset: 0x00000E74
         [HarmonyPostfix]
-        static void NightSanityModifier(ref decimal __result)
+        static void NightSanityModifier(ref float __result)
         {
             __result *= 2;
         }
@@ -29,7 +26,6 @@ namespace StressandTaxes.Costs.Patches
     [HarmonyPatch(typeof(GameConfigData), "GreaterMarrowDebtRepaymentProportion", MethodType.Getter)]
     public static class GameConfigData_GreaterMarrowDebtRepaymentProportion_Patch
     {
-        // Token: 0x06000007 RID: 7 RVA: 0x00002C74 File Offset: 0x00000E74
         [HarmonyPostfix]
         static void GreaterMarrowDebtRepaymentProportion(ref decimal __result)
         {
